Decompile scrcmd_pokemon_misc, 8

+++ include/unk_02031B0C.h
@@ -7,7 +7,7 @@ typedef struct SaveApricornBox SaveApricornBox;
 
 SaveApricornBox *Save_ApricornBox_get(SAVEDATA *saveData);
 void sub_02031CCC(SaveApricornBox *, int);
-int ApricornBox_CountApricorn(SaveApricornBox *apricornBox, u8 a1);
+int ApricornBox_CountApricorn(SaveApricornBox *apricornBox, u32 a1);
 void ApricornBox_GiveApricorn(SaveApricornBox *apricornBox, u16 a1, u8 a2);
 
 #endif //POKEHEARTGOLD_UNK_02031B0C_H
